Login is now required for all admin features (no accidental access) Improved promoting to subscription

+++ monosaur/views.py
@@ -13,6 +13,8 @@ from spend_analyser.transactions import transaction_handler
 
 
 def company(request):
+    if not request.user.is_superuser:
+        return redirect("/admin")
     CompaniesFormSet = modelformset_factory(Company, fields=['reference', 'name', 'category'], extra = 3)
 
     if request.method == "POST":
@@ -28,7 +30,9 @@ def company(request):
     return render(request, 'monosaur/companies.html', content)
 
 def subscription(request):
-    SubscriptionsFormSet = modelformset_factory(Subscription, fields='__all__', extra=3)
+    if not request.user.is_superuser:
+        return redirect("/admin?next=www.google.com")
+    SubscriptionsFormSet = modelformset_factory(Subscription, fields=['reference', 'company', 'name', 'description', 'monthly_price', 'subscription_url'], extra=2)
     
     if request.method == "POST":
         formset = SubscriptionsFormSet(request.POST or None)
@@ -43,23 +47,40 @@ def subscription(request):
     
 
 def db_cleanup(request):
+    if not request.user.is_superuser:
+        return redirect("/admin")
     transaction_handler.delete_old_entries()
     return redirect("/analyse")
 
 def db_clear(request):
+    if not request.user.is_superuser:
+        return redirect("/admin")
     transaction_handler.delete_all_entries()
     return redirect("/analyse")
 
 def delete(request, table, pk):
+    if not request.user.is_superuser:
+        return redirect("/admin")
     if request.user.is_superuser:
         if table=='company':
-            Company.objects.filter(pk=pk).delete()
+            deleted, row_count = Company.objects.filter(pk=pk).delete()
             Company.save_to_fixture()
         elif table=='subscription':
-            Subscription.objects.filter(pk=pk).delete()
+            deleted, row_count = Subscription.objects.filter(pk=pk).delete()
             Subscription.save_to_fixture()
     return redirect(request.GET.get('next', '/analyse'))
 
 def logout(request):
+    if not request.user.is_superuser:
+        return redirect("/admin")
     auth.logout(request)
+    return redirect(request.GET.get('next', '/analyse'))
+
+def promote_to_subscription(request, pk):
+    if not request.user.is_superuser:
+        return redirect("/admin")
+    company = Company.objects.get(pk=pk)
+    subscription = Subscription(company=company, reference=company.reference, name=company.name)
+    subscription.save()
+    print(str(subscription.pk))
     return redirect(request.GET.get('next', '/analyse'))
